refactor(server): rename commands
