+++ server/src/scripts/import-lovac/source-owners/source-owner-command.ts
@@ -1,12 +1,13 @@
+import fs from 'node:fs';
+import { Readable } from 'node:stream';
+
+import { countLines } from '@zerologementvacant/utils';
 import { createLogger } from '~/infra/logger';
 import { createLoggerReporter } from '~/scripts/import-lovac/infra';
 import {
   SourceOwner,
   sourceOwnerSchema
 } from '~/scripts/import-lovac/source-owners/source-owner';
-import { countLines } from '@zerologementvacant/utils';
-import { Readable } from 'node:stream';
-import fs from 'node:fs';
 import createSourceOwnerFileRepository from '~/scripts/import-lovac/source-owners/source-owner-file-repository';
 import { progress } from '~/scripts/import-lovac/infra/progress-bar';
 import validator from '~/scripts/import-lovac/infra/validator';
